backend/requirements.txt
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Django==3.2.13
django-filter==22.1
Pillow==9.1.1
gunicorn==20.0.4
PyJWT==2.4.0
psycopg2-binary==2.9.3
python-dotenv==1.0.0
drf-extra-fields==3.3.0
reportlab==4.0.4


